requirements.txt
Click>=6.0
dlib>=19.3.0
numpy
Pillow
scipy>=0.17.0
pandas
opencv-python
python-dotenv
supabase